Extract association business logic into AbstractSchemaRegistry
Move storage-agnostic association query, validation, and mutation logic
from KafkaSchemaRegistry to AbstractSchemaRegistry so that subclasses
can inherit shared behavior instead of duplicating it. Introduces four
protected hook methods for store-specific operations:
syncBeforeAssociationWrite, putAssociation, deleteAssociationEntry,
and lockForAssociation.
